更新购物车模型，调整字段类型为 uint32 和 int32，修正添加商品服务中的用户 ID 处理

+++ app/cart/biz/service/add_item.go
@@ -31,6 +31,6 @@ func (s *AddItemService) Run(req *cart.AddItemReq) (resp *cart.AddItemResp, err 
 	if req.UserId == 0 {
 		return nil, kerrors.NewGRPCBizStatusError(200401, "user id is required")
 	}
-	err = model.AddItem(s.ctx, mysql.DB, uint(req.UserId), model.Item{ProductId: uint(req.Item.ProductId), Quantity: int(req.Item.Quantity)})
+	err = model.AddItem(s.ctx, mysql.DB, req.UserId, model.Item{ProductId: req.Item.ProductId, Quantity: req.Item.Quantity})
 	return
 }
